+++ opentis/__init__.py
@@ -32,3 +32,5 @@ from visualize import PathTreeBuilder
 
 from volume import Volume, VolumeCombination, VolumeFactory, VoronoiVolume, \
     EmptyVolume, FullVolume, LambdaVolume, LambdaVolumePeriodic
+
+from todict import ObjectJSON, creatable, dictable, class_list
