+++ opentis/storage/__init__.py
@@ -1,8 +1,6 @@
-from netcdf_storage import Storage
+from storage import Storage, StorableObjectJSON
 from object_storage import ObjectStorage
-from ensemble_store import EnsembleStorage
 from snapshot_store import SnapshotStorage, ConfigurationStorage, MomentumStorage
 from trajectory_store import TrajectoryStorage
 from sample_store import SampleStorage, SampleSetStorage
-from engine_store import DynamicsEngineStorage
 from orderparameter_store import ObjectDictStorage
