+++ opentis/todict.py
@@ -1,13 +1,12 @@
-import json
-import yaml
 import base64
-import simtk.unit as units
+import json
 import mdtraj as md
 import numpy as np
 import pandas as pd
-import simtk.openmm.app
-
-import opentis.wrapper as wp
+from simtk import unit as units
+import simtk.openmm
+import yaml
+import opentis
 
 class ObjectJSON(object):
     """
@@ -20,7 +19,7 @@ class ObjectJSON(object):
         if class_list is not None:
             self.class_list = class_list
         else:
-            self.class_list = wp.class_list
+            self.class_list = opentis.todict.class_list
 
     def simplify_object(self, obj, base_type = ''):
         return { '_cls' : obj.__class__.__name__, '_dict' : self.simplify(obj.to_dict(), obj.base_cls_name) }
@@ -180,32 +179,76 @@ class ObjectJSON(object):
     def unit_from_json(self, json_string):
         return self.unit_from_dict(self.from_json(json_string))
 
+# Register a class to be creatable. Basically just a dict to match a classname to the actual class
+# This is mainly for security so that we do not have to use globals to find classes
 
-class StorableObjectJSON(ObjectJSON):
-    def __init__(self, storage, unit_system = None, class_list = None):
-        super(StorableObjectJSON, self).__init__(unit_system, class_list)
-        self.excluded_keys = ['name', 'idx', 'json', 'identifier']
-        self.storage = storage
+class_list = dict()
 
-    def simplify(self,obj, base_type = ''):
-        if type(obj).__module__ != '__builtin__':
-#            print obj.__dict__, hasattr(obj, 'creatable')
-            if hasattr(obj, 'idx') and (not hasattr(obj, 'nestable') or (obj.base_cls_name != base_type)):
-                # this also returns the base class name used for storage
-                # store objects only if they are not creatable. If so they will only be created in their
-                # top instance and we use the simplify from the super class ObjectJSON
-                base_cls = self.storage.save(obj)
-                return { '_idx' : obj.idx[self.storage], '_base' : base_cls, '_cls' : obj.__class__.__name__ }
+def creatable(super_class):
+    class_list[super_class.__name__] = super_class
 
-        return super(StorableObjectJSON, self).simplify(obj, base_type)
+    super_class.creatable = True
+    if not hasattr(super_class, '_excluded_attr'):
+        super_class._excluded_attr = []
 
-    def build(self,obj):
-        if type(obj) is dict:
-            if '_base' in obj and '_idx' in obj:
-                result = self.storage.load(obj['_base'], obj['_idx'])
-                # restore also the actual class name
+    if not hasattr(super_class, 'to_dict'):
+        def _to_dict(self):
+            excluded_keys = ['idx']
+            return {key: value for key, value in self.__dict__.iteritems() if key not in excluded_keys and key not in self._excluded_attr and not key.startswith('_')}
 
-                result.cls = obj['_cls']
-                return result
+        super_class.to_dict = _to_dict
 
-        return super(StorableObjectJSON, self).build(obj)
+    if not hasattr(super_class, 'from_dict'):
+        def _from_dict(cls, my_dict = None):
+            if my_dict is None:
+                my_dict={}
+
+            return cls(**my_dict)
+
+        super_class.from_dict = classmethod(_from_dict)
+
+    return super_class
+
+
+class LoadedObject(object):
+    pass
+
+
+def dictable(super_class):
+    """
+    A class decorator that marks a class to be storable in the storage using a LoadedObject class.
+    This object will have the same class name, the same dict, but none of the functions and will not have
+    been initialized. If you want real objects use @creatable
+    :param super_class: The class to be decorated
+    :return: The decorated class
+    """
+    class_list[super_class.__name__] = super_class
+
+    super_class.dictable = True
+    if not hasattr(super_class, '_excluded_attr'):
+        super_class._excluded_attr = []
+
+    if not hasattr(super_class, 'to_dict'):
+        def _to_dict(self):
+            excluded_keys = ['idx']
+            return {key: value for key, value in self.__dict__.iteritems() if key not in excluded_keys and key not in self._excluded_attr and not key.startswith('_')}
+
+        super_class.to_dict = _to_dict
+
+    if not hasattr(super_class, 'from_dict'):
+        def _from_dict(cls, my_dict = None):
+            if my_dict is None:
+                my_dict={}
+
+            obj = LoadedObject()
+
+            for key, value in my_dict.iteritems():
+                setattr(obj, key, value)
+
+            setattr(obj, 'cls', cls.__name__)
+
+            return obj
+
+        super_class.from_dict = classmethod(_from_dict)
+
+    return super_class
