+++ opentis/version.py
@@ -2,8 +2,8 @@
 # This file is automatically generated by setup.py
 short_version = '0.0.1'
 version = '0.0.1'
-full_version = '0.0.1.dev-f0c76c3'
-git_revision = 'f0c76c3515c73819a1ba67b3ff23408cde186f6d'
+full_version = '0.0.1.dev-58766f0'
+git_revision = '58766f06ed2b2e14af8a30662d751ba77d69a132'
 release = False
 
 if not release:
