+++ opentis/wrapper.py
@@ -56,74 +56,4 @@ def nestable(super_class):
     super_class.nestable = True
     return super_class
 
-# Register a class to be creatable. Basically just a dict to match a classname to the actual class
-# This is mainly for security so that we do not have to use globals to find classes
 
-class_list = dict()
-
-def creatable(super_class):
-    class_list[super_class.__name__] = super_class
-
-    super_class.creatable = True
-    if not hasattr(super_class, '_excluded_attr'):
-        super_class._excluded_attr = []
-
-    if not hasattr(super_class, 'to_dict'):
-        def _to_dict(self):
-            excluded_keys = ['idx']
-            return {key: value for key, value in self.__dict__.iteritems() if key not in excluded_keys and key not in self._excluded_attr and not key.startswith('_')}
-
-        super_class.to_dict = _to_dict
-
-    if not hasattr(super_class, 'from_dict'):
-        def _from_dict(cls, my_dict = None):
-            if my_dict is None:
-                my_dict={}
-
-            return cls(**my_dict)
-
-        super_class.from_dict = classmethod(_from_dict)
-
-    return super_class
-
-class LoadedObject(object):
-    pass
-
-def dictable(super_class):
-    """
-    A class decorator that marks a class to be storable in the storage using a LoadedObject class.
-    This object will have the same class name, the same dict, but none of the functions and will not have
-    been initialized. If you want real objects use @creatable
-    :param super_class: The class to be decorated
-    :return: The decorated class
-    """
-    class_list[super_class.__name__] = super_class
-
-    super_class.dictable = True
-    if not hasattr(super_class, '_excluded_attr'):
-        super_class._excluded_attr = []
-
-    if not hasattr(super_class, 'to_dict'):
-        def _to_dict(self):
-            excluded_keys = ['idx']
-            return {key: value for key, value in self.__dict__.iteritems() if key not in excluded_keys and key not in self._excluded_attr and not key.startswith('_')}
-
-        super_class.to_dict = _to_dict
-
-    if not hasattr(super_class, 'from_dict'):
-        def _from_dict(cls, my_dict = None):
-            if my_dict is None:
-                my_dict={}
-
-            obj = LoadedObject()
-
-            for key, value in my_dict.iteritems():
-                setattr(obj, key, value)
-
-            setattr(obj, 'cls', cls.__name__)
-
-            return obj
-
-        super_class.from_dict = classmethod(_from_dict)
-
-    return super_class
